# backend/tests/contracts/openapi_v0_1.json
{
  "components": {
    "schemas": {
...
      "ConsentDecision": {
        "properties": {
          "decision": {
            "enum": ["GRANTED", "WITHDRAWN"],
            "title": "Decision",
            "type": "string"
          },
          "purpose_code": {
            "enum": ["ACCOUNT", "AUDIO", "RESEARCH", "SECONDARY_USE"],
            "title": "Purpose Code",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": ["purpose_code", "version", "decision"],
        "title": "ConsentDecision",
        "type": "object"
      },
...
      "DeleteAccountRequest": {
        "properties": {
          "confirmation": {
            "const": "ELIMINAR MI CUENTA",
            "title": "Confirmation",
            "type": "string"
          }
        },
        "required": ["confirmation"],
        "title": "DeleteAccountRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "PWA Autonomía Económica Femenina",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/consents": {
      "post": {
        "operationId": "decide_consent_consents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Decide Consent Consents Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decide Consent",
        "tags": ["privacy"]
      }
    },
...
    "/privacy/deletion": {
      "post": {
        "operationId": "request_account_deletion_privacy_deletion_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Request Account Deletion Privacy Deletion Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Account Deletion",
        "tags": ["privacy"]
      }
    },
    "/privacy/export": {
      "post": {
        "operationId": "request_data_export_privacy_export_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Request Data Export Privacy Export Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Data Export",
        "tags": ["privacy"]
      }
    },
...
  }
}